python/packages/corlinman-agent/pyproject.toml
[project]
name = "corlinman-agent"
version = "0.1.0"
description = "Self-hosted reasoning loop, OpenAI tool_calls handling, context assembler, approval gate."
requires-python = ">=3.12"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "corlinman contributors" }]

dependencies = [
    "pydantic>=2.10",
    "structlog>=25.1",
    "PyYAML>=6.0",
    "numpy>=1.26",
    "corlinman-grpc",
    "corlinman-providers",
    "corlinman-tagmemo",
]

[tool.uv.sources]
corlinman-grpc = { workspace = true }
corlinman-providers = { workspace = true }
corlinman-tagmemo = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/corlinman_agent"]
